bots/discord_bot/requirements.txt
flask
discord.py
cloudscraper
pandas
requests
PyGithub
python-dotenv
